fix: corrige ubicacion de puntos Mcal en GeoJSON
- CRS declarado como objeto named (urn:ogc:def:crs:EPSG::32719) en lugar
  de string plano; QGIS ahora lee el sistema de referencia correctamente
- pixel_to_world usa col+0.5/row+0.5 para devolver centro del pixel
  en vez de esquina superior-izquierda (offset de 5m en pixels 10m)
- Actualiza anotacion de tipo en pixel_to_world: int a float

+++ src/satplatform/contracts/geo.py
@@ -135,7 +135,7 @@ def bounds_to_geotransform(bounds: Bounds, width: int, height: int) -> GeoTransf
     py = (miny - maxy) / float(height)  # negativo (origen en esquina sup-izq)
     return (minx, px, 0.0, maxy, 0.0, py)
 
-def pixel_to_world(col: int, row: int, gt: GeoTransform) -> Tuple[float, float]:
+def pixel_to_world(col: float, row: float, gt: GeoTransform) -> Tuple[float, float]:
     x0, px, rx, y0, ry, py = gt
     x = x0 + col * px + row * rx
     y = y0 + col * ry + row * py
